\name{Aspirin}
\alias{Aspirin}
\docType{data}
\title{ A contingency table for heart attacks and aspirin use. }
\description{
  The Physicians health study data cross-classified according to Infarct
  (heart attack or not) and Group (Placebo or Aspirin).
}
\usage{
data(Aspirin)
}
\format{
  A 2 by 2 matrix of counts with row names:
  
  Infarct:Yes and Infarct:No,
  
  and column names:
  
  Group:placebo and Group:aspirin.
  }
\source{
  Steering Committee of the Physicians' Health Study Research Group. 1989.
  Final report of the aspirin component of the ongoing Physicians' Health
  Study. N Engl J Med, 321:129--135.
  
  van Hulst, R. 2018. Evaluating Scientific Evidence. ms.
}
\keyword{datasets}