\name{binPriorsPlot}
\alias{binPriorsPlot}
\title{
function to plot diverse Beta distributions for use as Binomial priors
}
\description{
This function just plots some Beta distributions with commonly used parameters
}
\usage{
binPriorsPlot()
}
\value{
none produced, the function just produces one (compound) plot
}
\references{
van Hulst, R. 2018. Evaluating Scientific Evidence. ms.
}
\author{
Robert van Hulst
}
\examples{
binPriorsPlot()
}

\keyword{ distributions }
