\name{evidence-package}
\alias{evidence-package}
\alias{evidence}
\docType{package}
\title{evidence: Functions and Data for Bayesian and Likelihood Analysis
}
\description{
The functions in this package include Bayesian and likelihood
alternatives to the standard statistical hypothesis tests that form part
of base R. Their aim is to provide a wider perspective on how statistical
evidence can be analyzed than the usual hypothesis-testing one. In view
of the increasing importance in science of Bayesian and likelihood
inference a wider exposure to these alternatives has become overdue.

This package makes Bayesian and likelihood analyses of simple statistical
problems as convenient as traditional frequentist ones are in R. In
addition, it makes effective use of R's excellent plotting capabilities,
and facilitates exploratory data analysis and an interactive approach to
modeling. Both data exploration and model exploration are crucial in
data analysis, and these are facilitated by an interactive and
graphics-centered approach. 
}

\details{
\tabular{ll}{
Package: \tab evidence\cr
Type: \tab Package\cr
Version: \tab 0.8.10\cr
Date: \tab 2018-04-15\cr
License: \tab GPL\cr
}
}
\author{
Robert van Hulst

Maintainer: <rvhulst.ubishops.ca>

}
\references{
van Hulst, R. 2018. Evaluating Scientific Evidence. ms.
}
\keyword{ package }
