\name{rats}
\alias{rats}
\alias{rats}
\title{ Carcinogenesis data on rats exposed to a carcinogen. }
\description{
  Up to forty eight rats were exposed to the carcinogen retinyl acetate
  or to a placebo in their diet, after which the number of tumors they developed
  was evaluated.
}
\usage{data(rats)}
\format{
  A data frame with 71 observations on the following 2 variables:
  \describe{
    \item{\code{y}}{number of rats that developed tumors}
    \item{\code{N}}{number of rats in group}
  }
}
\source{
  Gail, M.H., Santner, T.J., and Brown, C.C. (1980) An analysis of
     comparative carcinogenesis experiments based on multiple times to
     tumor. Biometrics 36, 255--266.
  }
\references{
  van Hulst, R. 2018. Evaluating Scientific Evidence. ms.
}

\keyword{datasets}
