\name{Smoking}
\alias{Smoking}
\docType{data}
\title{ Mortality due to heart infarct in smokers and non-smokers. }
\description{
  The data are from a retrospective study that compared mortality due
  to a heart infarct in people who smoked and sex-matched controls who
  did not.
}
\usage{data(Smoking)}
\format{
  A matrix with 781 observations cross-classified on the following 2
  factors: ``Infarct'' (''Yes'' or   ``Control'', rows), and
  ``EverSmoked'' (''Yes'' or   ``No'', columns).
  }
\source{ unknown }
\references{ van Hulst, R. 2018. Evaluating Scientific Evidence. ms.
}

\keyword{datasets}
