\name{SSQ}
\alias{SSQ}
\title{
A support function that calculates the sum of squares of a data vector.
}
\description{
The sum of squares of the input vector is returned.
}
\usage{
SSQ(x)
}
\arguments{
  \item{x}{
a vector of numbers without missing values
}
}
\value{
the sum of squares of x
}
\author{
Robert van Hulst
}

\examples{
SSQ(x = rnorm(n=100))
}


